Firmware: Update for size optimization and alignment with dfw
Update structure size and align with dynamic Firmware tool for
next kernel.

+++ tools/firmware/abe_typ.h
@@ -421,8 +421,8 @@ struct omap_aess_dma_offset {
  */
 
 struct omap_aess_task {
-	int frame;
-	int slot;
+	u8 frame;
+	u8 slot;
 	u16 task;
 };
